backend/requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
djangorestframework-simplejwt
supabase
simpleeval
dj-database-url
